reorganize documents screen files
