+++ packages/client/src/components/screens/documents/all-documents/ documents-filters.tsx
@@ -5,13 +5,13 @@ import { Controller, SubmitHandler, useForm } from 'react-hook-form';
 import { Filter } from 'tabler-icons-react';
 import { ActionIcon, Indicator, MultiSelect, SimpleGrid, Switch } from '@mantine/core';
 import { DatePickerInput } from '@mantine/dates';
-import { DocumentsFilters as DocumentsFiltersType } from '../../gql/graphql.js';
-import { TimelessDateString } from '../../helpers/dates.js';
-import { isObjectEmpty, TIMELESS_DATE_REGEX } from '../../helpers/index.js';
-import { useGetFinancialEntities } from '../../hooks/use-get-financial-entities.js';
-import { useUrlQuery } from '../../hooks/use-url-query.js';
-import { UserContext } from '../../providers/user-provider.js';
-import { PopUpModal } from '../common/index.js';
+import { DocumentsFilters as DocumentsFiltersType } from '../../../../gql/graphql.js';
+import { TimelessDateString } from '../../../../helpers/dates.js';
+import { isObjectEmpty, TIMELESS_DATE_REGEX } from '../../../../helpers/index.js';
+import { useGetFinancialEntities } from '../../../../hooks/use-get-financial-entities.js';
+import { useUrlQuery } from '../../../../hooks/use-url-query.js';
+import { UserContext } from '../../../../providers/user-provider.jsx';
+import { PopUpModal } from '../../../common/index.js';
 
 interface DocumentsFiltersFormProps {
   filter: DocumentsFiltersType;
